Points remaining and how race distribution
Points remaining had a bug since it was hard coded into the placedPoints variable. Fixed by moving the remaining points into the HTML instead.

In order to make it easier to display the race bonuses in the table header the chosenRace observable/subject is now distributing IRace and not string.