Removed alot of console.logs
Styled the armory container a bit.

Fixed that the applied bonuses from weapons are being reset upon chosing a new weapon skill. This also applies for offhand when user place points into shield.

Fully implemented offhand selector along with bonus distribution for it.

Fixed the summarizeEachColumn-function in table to be a promise in order to select where to apply the setCurrentPoints-function.

Added a "constants"-folder to support and placed bonuses there.

Refactored some code in global.service